Resetting the canvas resets the background color
Resetting the PsyCanvas results in a reset of the background color,
which is [.5, .5, .5, 1.0] for r, g, b, a respecitively. The image
is an integer representation of with 8bit integer values, hence since
there is no value hence gray (128), is slightly more than .5 when
converted back to float, we allow a float comparison that allows
1/255 as value for epsilon. And then the test pass.
Since we want to know that the picture isn't upside down e.g., this
is good enough.