Add init function for psylib, bugfixes timers lifetime
Now psylib should be initialized with psy_init and deinitialized with
psy_deinit. This is a start for feature #124

As the timer thread is not start at the start of the program and stopped
at the end, the timer lifetime is less confusing. No more timers
are fired that should have been destroyed.