Fix regression in scheduling stimuli
When the timers where added to the PsyStimulus class in order to connect
the Stimulus::started and ::finished signals to the on- and offset of
the stimuli, an incorrect timestamp was used. The timestamp used was the
desired onset. However, this isn't necessary the onset of the stimulus.
This caused errors in the presentation of subsequent stimuli, as the
timestamp was not correct. Now the scheduled timestamp is used to start
the timers, and this will allow to present the next stimuli at the
appropriate moment relative to a preceding stimulus.