Add utility to build and test the default builds
The run_test.py utility is a utility that knows about 5 different
types of builds. gcc and clang, debug and release and a build with
-fsanitize=address,undefined enabled.

This allows to run the unit-tests and examples in order to see
if the tests and examples are happy with the current state of psylib

In addition it makes the command to test everything a bit easier.
from: meson devenv -C build meson test
to: ./run_test.py build
or ./run_test.py all to run all 5 tests.
Additionally, this allows to locally setup the 5 environments relatively
easily.