Use munit with sources instead of subproject
When using munit as subproject, clang builds currently fail due to
using of _Atomic, which is a C11 feature. However, the subproject is
build with c_std=c99, hence, it emits an error and the build crashes as
it uses --werror.
As source, it gets our flags which use C11 or newer