meson test give timers a good chance to perform
Previously, timers were occasionally accurate when run
alone. When running meson test, multiple test run in parallel.
This makes the CPU rather busy, So it might be the case that
the timers perform cannot respond very quickly in the main thread,
when running simultaneously with other tests. Hence, this
commit runs the timer-test not in parallel with other tests.