WIN32 Call TimeBeginPeriod(0) on psylib init
Sleep(1), will sleep at least 1 ms on windows, but in practice
it sleeps for much longer e.g. +/- 10 times as long. For timing
stuff accuratly, psylib needs to busyloop until that time arives.
So in order to save energy, psylib sleeps for 1 ms until the
event we are sleeping for is less then 1 ms away, and then it will
start busy looping, until the event-time.

The code in [class.ParallelTrigger] was fine, however, it overslept
because Sleep issue mentionned above. Now TimeBeginPeriod is called
when initializing psylib and TimeEndPeriod is called when deinitialiing
psylib

This fixes #137